requirements.txt
asgiref==3.8.1
attrs==24.2.0
cachetools==5.5.0
coverage==7.6.4
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.3
django-rql==4.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.1.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lark-parser==0.11.0
lib-rql==2.0.2
mccabe==0.7.0
packaging==24.1
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
simplejwt==2.0.1
six==1.16.0
sqlparse==0.5.1
typing==3.7.4.3
uritemplate==4.1.1
